ironfish-rust-nodejs: move the multisig structs to the `frost` submodule
