--- ironfish-rust-nodejs/src/structs/key_packages.rs
@@ -1,22 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
-
-use napi_derive::napi;
-
-#[napi(object)]
-pub struct IdentityKeyPackage {
-    pub identity: String,
-    pub key_package: String,
-}
-
-#[napi(object)]
-pub struct TrustedDealerKeyPackages {
-    pub proof_authorizing_key: String,
-    pub view_key: String,
-    pub incoming_view_key: String,
-    pub outgoing_view_key: String,
-    pub public_address: String,
-    pub key_packages: Vec<IdentityKeyPackage>,
-    pub public_key_package: String,
-}
